# openapi.yml
openapi: 3.0.3
info:
  title: SIGC - Sistema de Gestión de Citas Médicas
  description: API completa para gestión de especialidades, doctores, citas y usuarios
  version: 1.0.0
  contact:
    name: Soporte SIGC
    email: soporte@sigc.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
  - url: https://api.sigc.com
    description: Servidor de producción
...
paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      description: Autentica un usuario y retorna un token JWT válido por 24 horas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: paciente@sigc.com
              password: Paciente123456
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                message: Login exitoso
                token: eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InBhY2llbnRlQHNpZ2MuY29tIiwicm9sIjoiUEFDSUVOVEUiLCJzdWIiOiI1IiwiaWF0IjoxNzY0OTc4Mjc3LCJleHAiOjE3NjUwNjQ2Nzd9
                rol: PACIENTE
                idUsuario: 5
                email: paciente@sigc.com
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Autenticación
      summary: Registrar nuevo usuario
      description: Crea una nueva cuenta de usuario en el sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              nombre: Juan Pérez
              email: juan@sigc.com
              password: Juan123456
              dni: "12345678"
              telefono: "555000100"
              rol: PACIENTE
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroResponse'
              example:
                idUsuario: 5
                nombre: Juan Pérez
                email: juan@sigc.com
                mensaje: Usuario registrado exitosamente
        '400':
          description: Datos de entrada inválidos
        '409':
          description: El email ya está registrado
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido al hacer login
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: paciente@sigc.com
        password:
          type: string
          format: password
          example: Paciente123456
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login exitoso
        token:
          type: string
          description: Token JWT válido por 24 horas
        rol:
          type: string
          enum: [PACIENTE, DOCTOR, ADMIN]
          example: PACIENTE
        idUsuario:
          type: integer
          example: 5
        email:
          type: string
          example: paciente@sigc.com
    RegisterRequest:
      type: object
      required:
        - nombre
        - email
        - password
        - dni
        - telefono
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
          example: Juan Pérez
        email:
          type: string
          format: email
          example: juan@sigc.com
        password:
          type: string
          minLength: 6
          maxLength: 50
          example: Juan123456
        dni:
          type: string
          pattern: '^\d{8}$'
          example: "12345678"
        telefono:
          type: string
          pattern: '^\d{9}$'
          example: "555000100"
        rol:
          type: string
          enum: [PACIENTE, DOCTOR, ADMIN]
          default: PACIENTE
          example: PACIENTE
    RegistroResponse:
      type: object
      properties:
        idUsuario:
          type: integer
        nombre:
          type: string
        email:
          type: string
        mensaje:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Credenciales inválidas
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 401